requirements.txt
asgiref==3.3.4
Django==3.2
django-rest-enumfield==0.2.0
django-rest-framework==0.1.0
djangorestframework==3.12.4
pytz==2021.1
sqlparse==0.4.1
unittest-xml-reporting==3.0.4
xmlrunner==1.7.7
yapf==0.31.0
asgiref==3.3.4
boto3==1.14.63
botocore==1.17.63
ca.authenticator==1.59.0
ca.django-rest-framework-jwt==3.2.0
ca.logs==0.13.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==2.9.2
Django==3.2
django-dotenv==1.4.2
django-health-check==3.14.0
django-rest-enumfield==0.2.0
django-rest-framework==0.1.0
django-storages==1.10.1
djangorestframework==3.12.4
docutils==0.15.2
elastic-apm==5.10.0
greenlet==1.1.0
idna==2.10
jmespath==0.10.0
psutil==5.8.0
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
rollbar==0.15.1
s3transfer==0.3.6
six==1.15.0
SQLAlchemy==1.4.15
sqlparse==0.4.1
unittest-xml-reporting==3.0.4
urllib3==1.25.11
uWSGI==2.0.19.1
xmlrunner==1.7.7
yapf==0.31.0
